update ( modal on, off add / quickbar add) fix ( styled name fix)
